yama: (concept) restrict seccomp-eBPF with ptrace_scope
LSM hook seccomp_extended is made to return -EPERM if the current
process may not ptrace its children, depending on the value of
ptrace_scope and CAP_SYS_PTRACE capability.

The hook checks CAP_SYS_PTRACE against the init userns if
ptrace_scope has a value of YAMA_SCOPE_CAPABILITY. It is because
an eBPF filter cannot be loaded in another userns. [Jinghao Jia]

I'm not sure if this is the right way to do it, since ptrace_scope
is about ptrace and not seccomp. Is there a better policy that would
make more sense?

Signed-off-by: YiFei Zhu <yifeifz2@illinois.edu>
Co-developed-by: Jinghao Jia <jinghao7@illinois.edu>
Signed-off-by: Jinghao Jia <jinghao7@illinois.edu>